[pipelining] RNG seed management for PP init
Implement both 'serial' and 'parallel' options described in #139304

'parallel' mode sets a unique seed per rank, but does not worry about
parity with non-PP initialization.  It is fastest becuase it
parallelizes initialization and uses no communication.

'serial' mode does initialization one stage at a time, communicating the
RNG states from rank to rank as needed so that the resulting
initialization is equivalent to non-PP initialization where the whole
model is initialized on one rank.

The overall initialization is delegated to the 'schedule' since it knows which
stages it owns, which ranks they are on, and what order they go in.
But the per-stage initialization is handled by a user-passed function
that can do anything it wants to the given module owned by that stage.

TODO: add tests and see if this actually works. See whether 'serial'
mode is slow for large PP setups.

ghstack-source-id: 83f0caefb724284096074e9e4f30e3fb6082bf98
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139445